[refactor-home] Refactored the Helpers object to be for ObligationStatus
