+++ app/controllers/ObligationsController.scala
@@ -25,7 +25,6 @@ import play.api.libs.json.JsResultException
 import play.api.mvc.{Action, AnyContent}
 import services.ObligationsService
 import uk.gov.hmrc.play.http.InternalServerException
-import views.Helpers
 
 import scala.concurrent.Future
